third_party/WebKit/LayoutTests/fast/events/touch/touch-user-gesture-expected.txt

CONSOLE WARNING: line 13: Performing operations that require explicit user interaction on touchstart events is deprecated and will be removed in M55, around December 2016. See https://www.chromestatus.com/features/5649871251963904 for more details.
CONSOLE WARNING: line 13: Performing operations that require explicit user interaction on touchmove events is deprecated and will be removed in M55, around December 2016. See https://www.chromestatus.com/features/5649871251963904 for more details.
Target
Test user gesture behavior during touch events.
touchstart should be a user gesture
PASS openedPopup is true
touchend for the same gesture should not get a new gesture
PASS openedPopup is false
touchmove should be a user gesture
PASS openedPopup is true
second touchmove for the same gesture should not get a new gesture
PASS openedPopup is false
touchend should be a user gesture
PASS openedPopup is true
PASS successfullyParsed is true

TEST COMPLETE
